added cpp files for player and computer class definitions
