+++ tic-tac-toe/x64/Debug/tic-tac-toe.tlog/Cl.items.tlog
@@ -1,2 +1,5 @@
+E:\Programming Projects\C++\tic-tac-toe\common.cpp;E:\Programming Projects\C++\tic-tac-toe\tic-tac-toe\x64\Debug\common.obj
+E:\Programming Projects\C++\tic-tac-toe\comp_player.cpp;E:\Programming Projects\C++\tic-tac-toe\tic-tac-toe\x64\Debug\comp_player.obj
 E:\Programming Projects\C++\tic-tac-toe\grid.h;E:\Programming Projects\C++\tic-tac-toe\tic-tac-toe\x64\Debug\grid.obj
+E:\Programming Projects\C++\tic-tac-toe\human_player.cpp;E:\Programming Projects\C++\tic-tac-toe\tic-tac-toe\x64\Debug\human_player.obj
 E:\Programming Projects\C++\tic-tac-toe\main.cpp;E:\Programming Projects\C++\tic-tac-toe\tic-tac-toe\x64\Debug\main.obj
