+++ tic-tac-toe/x64/Debug/tic-tac-toe.tlog/link.secondary.1.tlog
@@ -1,2 +1,2 @@
-^E:\PROGRAMMING PROJECTS\C++\TIC-TAC-TOE\TIC-TAC-TOE\X64\DEBUG\GRID.OBJ|E:\PROGRAMMING PROJECTS\C++\TIC-TAC-TOE\TIC-TAC-TOE\X64\DEBUG\MAIN.OBJ
+^E:\PROGRAMMING PROJECTS\C++\TIC-TAC-TOE\TIC-TAC-TOE\X64\DEBUG\COMMON.OBJ|E:\PROGRAMMING PROJECTS\C++\TIC-TAC-TOE\TIC-TAC-TOE\X64\DEBUG\COMP_PLAYER.OBJ|E:\PROGRAMMING PROJECTS\C++\TIC-TAC-TOE\TIC-TAC-TOE\X64\DEBUG\GRID.OBJ|E:\PROGRAMMING PROJECTS\C++\TIC-TAC-TOE\TIC-TAC-TOE\X64\DEBUG\HUMAN_PLAYER.OBJ|E:\PROGRAMMING PROJECTS\C++\TIC-TAC-TOE\TIC-TAC-TOE\X64\DEBUG\MAIN.OBJ
 E:\Programming Projects\C++\tic-tac-toe\tic-tac-toe\x64\Debug\tic-tac-toe.ilk
